PouchDB: Use listener pattern
Rather than manually fetching data only when the local user updates it,
this sets it up so that the local user listens to _any_ database change.
This means that even if someone _else_ updates the database, and then
the change comes in from the backend rather than being generated by this
user, this user will get it right away.

Note that we also unsubscribe from changes in the `componentWillUnmount`
hook. This probably isn't necessary at present, since as long as this
page is loaded, the App component will be loaded. But this may be
necessary down the road, and it's standard practice in React--If you
subscribe to something in `componentWillMount`, you need to unsubscribe
in `componentWillUnmount`.